+++ js/directives/pwDevices.js
@@ -49,3 +49,118 @@ postworld.directive('pwDeviceClass', function( $pw, $log ) {
 		}
 	}
 });
+
+
+/**
+ * @ngdoc directive
+ * @name postworld.directive:pwImage
+ * @description
+ * Shows the correct image based on the current device pixel ratio.
+ *
+ * @param Object pwImageSrc An object
+ *
+ * @example
+ * 		<img pw-image-src="{'1':'image.jpg','2':'image-hi-res.jpg'}">
+ *
+ */
+postworld.directive('pwImageSrc',
+	[ '$pw', '$log', '_', '$window', '$filter',
+	function( $pw, $log, $_, $window, $filter ) {
+	return {
+		restrict:'A',
+		scope:{
+			pwImageSrc:'='
+		},
+		link: function( $scope, element, attrs ) {
+			var srcs = $scope.pwImageSrc;// $scope.$eval( $scope.pwImageSrc );
+
+			if( _.isEmpty(srcs) )
+				return false;
+
+			//$log.debug( 'pwImageSrc', srcs );
+			//$log.debug( 'devicePixelRatio', $window.devicePixelRatio );
+
+			/**
+			 * Reconstruct the input sources
+			 * As an array of objects.
+			 */
+			var orderSrcs = function( srcs ){
+				// Placeholder array
+				var srcsArr = [];
+				// Construct array
+				angular.forEach( srcs, function( value, key ){
+					srcsArr.push({
+						pixelRatio: parseFloat(key),
+						src: value
+					});
+				});
+				// Order the array by the pixel ratio
+				srcsArr = $filter('orderBy')( srcsArr, 'pixelRatio', true );
+				return srcsArr;
+			}
+
+			/**
+			 * Using the current device's pixel ratio
+			 * get the appropriate image source.
+			 */
+			var selectSrc = function( options ){
+				// Placeholder for the chosen source
+				var src = '';
+				// The current device's pixel ratio
+				var devicePixelRatio = $window.devicePixelRatio;				
+				// Iterate through source options
+				angular.forEach( options, function( option ){
+					/**
+					 * If the option's ratio is greater than
+					 * Or equal to the current device pixel ratio
+					 * Then go ahead and use that ratio.
+					 */ 
+					if( option.pixelRatio >= devicePixelRatio  )
+						src = option.src;
+				});
+				/**
+				 * If no source has been found because the device
+				 * has a larger pixel ratio than what is available
+				 * then select the highest available ratio.
+				 */
+				if( _.isEmpty( src ) )
+					src = options[0].src;
+				return src;
+			}
+
+			/**
+			 * Set the actual element's src value.
+			 */
+			var setImageSrc = function(){
+				var src = selectSrc( orderSrcs( srcs ) );
+				src = $pw.injectVariables(src);
+				element.attr( 'src', src );
+			}
+
+			setImageSrc();
+
+			/**
+			 * If the src-update attribute is present
+			 * Watch for a change of device pixel ratio
+			 * Such as in a zoom in, and update the source.
+			 */
+			/*
+			if( !_.isUndefined( attrs.srcUpdate ) )
+				angular.element($window).bind("resize", elemClasses);
+
+				$scope.$watch(
+					// ISSUE : This isn't parsing?
+					function(){ return $window.devicePixelRatio; },
+					function( val ){
+						$log.debug( 'changing' );
+						setImageSrc();
+					}
+				);
+			else
+				setImageSrc();
+			*/
+				
+
+		}
+	}
+}]);