+++ php/core/styles.php
@@ -47,13 +47,14 @@ function pw_less_vars( $vars, $handle ) {
     ///// Directory Paths /////
     global $i_paths;
     global $pwSiteGlobals;
+    
+    // DEPRECIATED
     $vars['infinite-theme'] = pw_less_prepare_url( $i_paths['infinite']['url'] );
     $vars['child-theme'] = pw_less_prepare_url( $i_paths['child_theme']['url'] ); 
 
+    $vars['template-url'] = pw_less_prepare_url( get_template_directory_uri() ); 
+    $vars['theme-url'] = $vars['template-url'];
     $vars['postworld-url'] = pw_less_prepare_url( _get( $pwSiteGlobals, 'paths.postworld.url' ) );
-
-    $vars['i-templates-override'] = pw_less_prepare_url( $i_paths['templates']['url']['override'] );
-    $vars['i-templates-default'] = pw_less_prepare_url( $i_paths['templates']['url']['default'] );
 
     ///// CACHE /////
     $phpLessVarsCache = $vars;