+++ php/core/utilities.php
@@ -1196,6 +1196,20 @@ function pw_body_classes(){
 	return $body_classes;
 }
 
+function pw_html_classes(){
+	// Returns a string with the Wordpress body classes
+	$classes = array();
+	$output = '';
+
+	if( pw_module_enabled('devices') )
+		$classes = array_merge( $classes, pw_device_classes() );
+	
+	foreach( $classes as $class ){
+		$output .= " " . $class;
+	}
+	return $output;
+}
+
 
 function pw_get_menus(){
 	$menus = get_terms( 'nav_menu' );
@@ -1469,7 +1483,7 @@ function pw_reset_less_php_cache(){
 	//global $pwGlobalsJsFile;
 	$ghost_less_file = POSTWORLD_PATH .'/less/ghost.less';
 	$file = fopen( $ghost_less_file ,"w" );
-	fwrite($file,"// Reset PHP LESS Cache : " . date("Y-m-d H:i:s"));
+	fwrite($file,"// Reset PHP LESS Cache"); // . date("Y-m-d H:i:s"));
 	fclose($file);
 	//if( file_exists( $pwGlobalsJsFile ) )
 	//	chmod($pwGlobalsJsFile, 0755);